chore(revert): undo 'state' input layer and subsequent commits (revert cea989d..HEAD)

+++ src/int/sound.cc
@@ -341,8 +341,7 @@ static void refreshSoundBuffers(Sound* sound)
         } else {
             int bytesToRead = sound->dataSize;
             if (sound->field_58 != -1) {
-                /* compat_tell returns long - store in long to avoid narrowing */
-                long pos = sound->io.tell(sound->io.fd);
+                int pos = sound->io.tell(sound->io.fd);
                 if (bytesToRead + pos > sound->field_58) {
                     bytesToRead = sound->field_58 - pos;
                 }
@@ -382,8 +381,7 @@ static void refreshSoundBuffers(Sound* sound)
                         if (sound->field_58 == -1) {
                             bytesToRead = sound->dataSize - bytesRead;
                         } else {
-                            /* compat_tell returns long - store in long to avoid narrowing */
-                            long pos = sound->io.tell(sound->io.fd);
+                            int pos = sound->io.tell(sound->io.fd);
                             if (sound->dataSize + bytesRead + pos <= sound->field_58) {
                                 bytesToRead = sound->dataSize - bytesRead;
                             } else {
